Moved local storage & added Json methods
