+++ src/utils/local-storage.ts
@@ -1,0 +1,28 @@
+type StorageKey = "apiToken" | "userName" | JsonStorageKey;
+type JsonStorageKey = "goals";
+
+// TODO: #12 Now the compiler will be fine with `JsonStorageKey` when I'm trying
+//       to read a string key.
+//       Let's abstract a generic function:
+function getStringKey(key: StorageKey): string {
+  return localStorage.getItem(key) ?? ''
+}
+
+function storeStringKey(key: StorageKey, value: string) {
+  localStorage.setItem(key, value)
+}
+
+function getJsonKey(key: JsonStorageKey): any {
+  try {
+    return JSON.parse(getStringKey(key))
+  } catch {
+    return null;
+  }
+}
+
+function setJsonKey(key: JsonStorageKey, value: Object) {
+  storeStringKey(key, value.toString())
+}
+
+export { getStringKey, storeStringKey, getJsonKey, setJsonKey };
+
